+++ checkpoint/CHANGELOG.md
@@ -17,6 +17,8 @@ a `PathAwaitingCreation`.
   TypeHandlerRegistry.
 - #v1 Add tests to cover v0 CheckpointManager->v1 Checkpointer migration use
  case.
+- #v1 Add NumpyLeafHandler.
+- #v1 Add NumpyLeafHandler into CompatibleTypeHandler.
 
 ### Fixed